Update patch set 1
Patch Set 1:

According to the crbug, the problem is related to the voltage threshold:

"Vijay and I debugged this further and the issue seems to reside in charger where vbus level is determined without checking against VBUS detect threshold that is recommended (3900mV). With the proposed change the fix seems to be promising. Sent out image for more testing."

If the problem is actually due to voltage ripples < 3.7V then the solution here is not guaranteed to work (what if you read the analog voltage on a ripple?) and is probably the same as reading the PRESENT register twice.

If you're trying to fix both problems, there's a smoothed ADC sample reg VBUS_AVE_VAL/VCC_AVE_VAL that may help, though I'm not sure about the time sample paramters it uses (maybe ask Rohm).

Patch-set: 1