charge_state_v2: Add default value to charger power threshold to boot
Added default value of 15mW for CONFIG_CHARGER_LIMIT_POWER_THRESH_CHG_MW
config which is the minimum power needed by the external charger to boot
the system to OS when the battery is dead or not present. Override it in
board.h if the value is board specific.

BUG=chrome-os-partner:61801
BRANCH=none
TEST=Manually tested on Reef. Without the battery system is
     'power-up inhibited' with 15W charger.

Change-Id: I9e0b808a1b65948249f14ee361c3925050a0de02
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>