g: fix pinmux command implementation
The output of pinmux command is garbled (it probably was OK until the
console UART TX buffer was significantly reduced in size by
https://chromium-review.googlesource.com/418948).

Let's flush the console every so often.

Also, fixed the bug where the second GPIO port base address in PIMUX
was off, and added a newline to separate UART section of the output.

BRANCH=none
BUG=none
TEST=ran pinmux command successfully. make -j buildall passes

Change-Id: Iae0dcbbd0a55d704ffe2296d6e0e81b2c6aec527
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>