g: USB: do not disable sleep indefinitely on initialization
USB initialization routine currently disables sleep, leaving it to be
re-enabled by the ISR when the connection is suspended.

The problem is that there is no guarantee that the connection would be
established and then suspended soon after boot.

Let's not disable sleep after initialization. It will be disabled very
soon if USB connection is up.

BRANCH=none
BUG=chrome-os-partner:62348
TEST=verified successful Suzy-Q and internal USB ports operation.

Change-Id: I31b3991aed0fd71971b8c65d658dd50f869e7ede
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>