Reef: Set RW boot power threshold to 18w
Electro is able to boot on 15w to Depthcharge but requires 18w or above
to boot to the logon screen. This patch allows 15w charger to boot Electro
to Depthcharge (as before) but prevents it from booting after that unless
the EC in RW negotiates with the charger and gets 18w or more.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=chrome-os-partner:61801
BRANCH=none
TEST=Hack the code to enforce syslock. Disconnect battery.
Use 18w charger to boot the system. Use 15w only charger not to boot.

Change-Id: I00d79a96221f1d3b8c6d529de9b8e4588d6112aa