Update patch set 1
Patch Set 1:

Agree, it would be nice if we could commit an initial version, even if it's not perfect, so we can make life a bit easier for AP FW development (i.e. not having everybody do manual source patching of EC src to be able to build FW/coreboot).

If the only component we're uncertain about is the BMI160, can we
1. commit up until CL:427565 (WIP: rowan: enable
   CONFIG_CHARGE_MANAGER_EXTERNAL_POWER_LIMIT)
2. remove WIP in all CL leading up to that as suggested by Rong
3. create a new bug that is just adding proper support for BMI160 and
   just disable the accel completely?
   I don't know enough about this, but to disable would it be enough/viable
   to just comment out these lines
   #define CONFIG_ACCEL_KX022
   #define CONFIG_CMD_ACCELS
   #define CONFIG_CMD_ACCEL_INFO

Patch-set: 1