Update patch set 17
Patch Set 17:

> Patch Set 17: Code-Review-1
> 
> > Patch Set 17:
> > 
> > > Patch Set 17:
> > > 
> > > (3 comments)
> > 
> > Please help review this patch.
> 
> Did you test the patch against previous issues? It seems with this patch,  analogix port can no longer handle role swap properly. I see hard reset after SRC_SWAP_STANDBY->SNK_DISCOVERY. This is similar to Issue 62270 but I can reproduce it with the oem charger.

So are you testing with Apple dongle HDMI and type-c charger described in 62270. I do not have OEM charger, but during the sequence of state transition when connecting zinger after connecting HDMI, the state transitions from st6 to st2 but port does not go to low power mode?
Is it not still a separate issue?

Patch-set: 17