Update patch set 1
Patch Set 1: Code-Review-2

This is only half the issue. 

Another issue is that we set device_states[DEVICE_AP].detect to be GPIO_TPM_RST_L. Then in board_update_device_state we pass in the detect gpio to device_state_on if we dont already think the device is on.
device_state_on then immediately disables the interrupt on that gpio. This means the interrupt on GPIO_TPM_RST_L is disabled and we never see a tpm reset.

Patch-set: 1
Label: Code-Review=-2