Update patch set 2
Patch Set 2:

> Could we add a software flag to indicate the first byte is already overridden? Set it in IBF ISR and clear it after glitch detection. If this flag is set and the first byte is VERSION, return 0 directly.

The danger is that 0x3 may appear anywhere in the input, not just as the initial protocol byte (but it's uncommon, because the host will typically never output it, after sending the header + payload). So if we return 0 directly without waiting, there is some chance we will return too early and it will lead to a new failure.

Another idea is to check byte 0 at CS deassert, if it's non-0x3 then we can apply your suggestion. If it's 0x3, then maybe we need to wait out the entire duration.

Patch-set: 2
Label: Code-Review=0