cr50: remove encapsulated tpm commands
This removes the ability to call arbitrary tpm vendor commands
from cr50, and replcaes with upgrade-specific functionality.

NOTE: this is demo code, I didn't increase the version number
of the interface, and left some ifdefs. It's pretty straightforward
to clean up.

NOTE: I didn't implement "corrupt inactive RW", as I'm not sure
how to test it, however this is pretty much the sane, pull out
the functionality into a function called form the vendor command,
and make an updgrade command to do this, which calls the function.

BUG=chrome-os-partner:62815
BRANCH=None
TEST=updates just fine

Change-Id: I4be44362d6897ac8b39e134ba03b0047d8c0e95a