Update patch set 2
Patch Set 2:

I am not sure I am fully comfortable with this.

I don't think the counter should be reset to zero, I think it should be decremented if it is not zero. so that the upcoming increment by the RO is compensated.

Also, instead of adding a yet another system wide type why not just call a function 'decrement counter' from these reboot invocation points?

Patch-set: 2
CC: Gerrit User 1000229 <1000229@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>