Update patch set 1
Patch Set 1:

yeah, I just noticed this because I use an emacs macro which injects the guards all the way to the directory where .git is.

I think this should be the default, but agree that this is not a big deal.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>