bob: discharge when battery is near full
This is copy from CL:408735

For device that are using ceramic caps, when battery is full, discharge
to 97% and then charge again until the battery is full. This should also
help us meet CEC energy spec.

BUG=chrome-os-partner:62916
BRANCH=gru
TEST=Manual on bob, insert charger with battery at 97%, verify battery
does charge until 100%, and then discharge down to 96% and verify
battery charges. Charge to 100% again.

Change-Id: I6031a12c9f4e68cc75937861dfe81bdd7cb03df7
Signed-off-by: james_chao <james_chao@asus.com>