Update patch set 4
Patch Set 4:

(1 comment)

Reworked a lot more, to prevent size increases if the option is not defined.

When option is not defined, common/flash.c and stm32f0 code is a lot more similar to what it used to be before this change.

This still increases size by ~10 bytes on average across all boards, but we save about 18 bytes with the previous commit so I don't feel _too_ guilty.

Patch-set: 4
Label: Code-Review=0