Allow to optimize C routines using clang
Half-hackish build system to generate optimized assembly of speed
critical routines using LLVM/clang.
In the chroot, the LLVM toolchain has no bare-metal or soft-float versions
and the ARM-targeted compiler is not installed for all overlays.
Thus, we are not fully ready to build the full firmware with LLVM (and
we would need to fix so details in the current code).
So for now proceed in 3 steps:
- declare the objects we want to optimize in the 'use_opt_asm' build variable.
  (for example in the board build.mk file)
- using clang, generate (offline) optimized assembly files called
  <source>.<CORE>-genasm.S, by invoking 'make BOARD=<board> genopt'
  Those .S can be submitted into the source tree.
- during the firmware build, the selected objects are built by GCC/AS
  from those assembly files rather than doing the normal C file compilation.

Using the available LLVM 4.0ish in the chroot, I'm getting the following
improvements on the cortex-M4 based STM32L4:
- curve25519 crypto: 24.4% faster at 16Mhz (12.2% at 80Mhz)
 (for +960 bytes of code space)
- RSA crypto: 38.4% faster at 16Mhz (36.2% at 80Mhz)
 (for +216 bytes of code space)

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=chrome-os-partner:62991
TEST=manual
write a small cryptography benchmark (cf follow-up CL)
make BOARD=eve_fp genopt
make BOARD=eve_fp
run it on STM32L432

Change-Id: I9f7e0b8e9d846a2ecbfb8290ad420fe5226b7ac9