nvmem: do not prefer old partition over new one.
The code deciding which partition to chose properly determines which
one is newer, but invariably picks the second as the current, if both
partitions are valid.

This is obviously wrong, on top of that there is no need to check both
partitions' validity, if the newer one checks out.

BRANCH=none
BUG=chrome-os-partner:60555, chrome-os-partner:61972
BUG=chrome-os-partner:61716, chrome-os-partner:61234
BUG=chrome-os-partner:61167, chrome-os-partner:60555
TEST=successfully ran
   $ ssh 192.168.1.102 suspend_stress_test --suspend_min 10 \
         --suspend_max 10 --wake_max 10 --wake_min 10
   for more than 300 cycles (it still is going)

Change-Id: I02d9bb062b9edbbb9787a95ba760872e09ff2761
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>