Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (2 comments)

guys, this is not solving all problems and leaves a couple of corner cases open, but this definitely is an improvement.

Does anyone think this patch needs to change. if so - how? If not, can I please get a +2! :)

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>