cr50: read fwmp and act on it when controlling console restrictions
It needs to be possible to prevent unlocking of CCD on enterprise
enrolled devices, in particular to prevent users to moving into dev
mode.

A bit in the FWMP structure flags field was allocated for the purposes
of preventing console unlock in those cases.

This patch adds code to read the FWMP structure from the TPM NVMEM,
verify it and determine if it should be possible to unlock the
console.

The FWMP value is read only once on TPM reset, this means once the
admin console changes the relevant flag bit, the Chrome OS device will
have to be rebooted to pick up the new flag value.

BRANCH=cr50
BUG=b:35587387,b:35587053
TEST=not much yet, waiting for the upper layers to allow configuring
     FWPM

Change-Id: I17e15ea2b2293a0c096858fba3ccc389452caede