cr50: read fwmp and act on it
It needs to be possible to prevent unlocking of CCD on enterprise
enrolled devices, to prevent users to moving into dev mode.

A bit in the FWMP structure flags field was allocated for the purposes
of preventing console unlock in those cases.

This patch adds code to read the FWMP structure from the TPM NVMEM,
verify it and determine if it should be possible to unlock the
console.

The function reporting the console restriction state could be called
on different contexts, sometimes with a limited stack size. NVMEM
access functions are fairly stack heavy, so to avoid increasing the
console stack size let's move the check to the hooks task context.

BRANCH=cr50
BUG=b:35587387,b:35587053
TEST=not much yet, waiting for the upper layers to allow configuring
     FWPM

Change-Id: I17e15ea2b2293a0c096858fba3ccc389452caede