cr50: read fwmp and act on it when controlling console restrictions
It needs to be possible to prevent unlocking of CCD on enterprise
enrolled devices, in particular to prevent users from moving into dev
mode.

A bit in the FWMP structure flags field was allocated for the purposes
of preventing console unlock in those cases.

This patch adds code to read the FWMP structure from the TPM NVMEM,
verify it and determine if it should be possible to unlock the
console. The restriction is not honored by Cr50 DBG images.

The FWMP value is read only once per TPM reset, this means each time
the admin console changes the relevant flag bit, the Chrome OS device
has to be rebooted to pick up the new flag value.

BRANCH=cr50
BUG=b:35587387,b:35587053
TEST=verified that FWMP is properly read and acted upon.

Change-Id: I3de84b23b4f2ec5b5d8d9a2ebbdb27ed384a7eea
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: fb5a05ab223b56b22a3e0978333586ec13604374
Original-Change-Id: I17e15ea2b2293a0c096858fba3ccc389452caede
Original-Reviewed-on: https://chromium-review.googlesource.com/457824
Original-Commit-Ready: Vadim Bendebury <vbendeb@chromium.org>
Original-Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Original-Reviewed-by: Mary Ruthven <mruthven@chromium.org>