Update patch set 4
Patch Set 4:

We may have gotten ambitious when we added support for the new rwsig format.  This is currently the *only* thing which uses the vboot 2.1 data structures, which were targeted at vboot_reference and are likely overkill for the EC (though the old totally-bare uspbd1 format was underkill...)

As I remember, we crammed rwsig into the code in a hurry to support hadoken.  In retrospect a more compact header would have been better, both in binary size and in the code required to parse it in the EC.

Patch-set: 4
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>