pd: Add tcpc_select_rp_value to stub tcpm driver
The function tcpm_select_rp_value() is used to set a given Rp
value. This function was not supported for boards that use the config
option CONFIG_USB_PD_TCPC. This CL adds a function call for systems
that have CONFIG_USB_PD_DTS defined. The implementation of this
function is board specific and would be in the file
/board/<>/usb_pd_config.h file.

BUG=chrome-os-partner:61878
BRANCH=servo
TEST=run 'make buildall'.

Change-Id: I1cd6f0ed51717736d30575c58025bd9a1c5d6fc4
Signed-off-by: Scott <scollyer@chromium.org>