servo_v4: (WIP)Enable VBUS passthrough from CHG to DUT port
This CL adds support for the DUT port to advertise different current
levels via the Rp resistor selection. The default setting is USB (500
mA). Added a board_manage_port function that is called whenever VBUS
is turned on or when input current limit is set. When
board_manage_port is called by the CHG port it may trigger an update
to the DUT VBUS source.

Added an implementation of tcpm_select_rp_value for devices such as
servo_v4 which run tcpc functions on the EC.

The VBUS passthrough mode assumes (but also checks) that CHG VBUS is
not greater than 5V since the DUT port still doesn't not have the
ability to send PD messages and establish an explicit contract.

BUG=chrome-os-partner:61878
BRANCH=servo
TEST=Manual
With CHG port disconnected, connected DUT to Reef, verified that SRC
accessory mode was enabled and that on Reef side it went into
SNK_ACCESSORY state. Then connected Zinger to CHG port, verified that
there was a detach event, and when reconnecting DUT port used CHG
passthrough (the LED is green instead of amber). Disconnected Zinger
and verified the detach on DUT port and that it reconnected using the
host VBUS.

Change-Id: Ib7491be5513fe3fbbf7e6cd084b59fd4cf346ab6
Signed-off-by: Scott <scollyer@chromium.org>