Update patch set 5
Patch Set 5:

(1 comment)

> What this means is that we definitely can't do normal development with recovery mode w/o a battery. This policy might have an impact on the forthcoming chromebox.

We're probably going to negotiate PD in RO for the devices you mentioned. That means we can negotiate back to 5V. But we can't do that here -- if we suddenly find ourselves in RO with PD previously negotiated, what do we do? Either interrupt the contract (as we're doing here) or don't (and we're back to the old inrush problem that we're trying to fix).

Patch-set: 5