Added platform/ec/chip/g/trng_cprng.c.
This initializes an AES based CPRNG from the TRNG during init_trng().
It then runs a simple health check to determine that the LSB of each
TRNG sample is not well correlated with the prior 2 sample LSBs.  The
original trng.c remains for use in ght RO code.  The resulting CPRNG
output should be more unpredictable and significantly faster.

BRANCH=none
BUG=none
TEST=none

Change-Id: I936e81dc5eae05286677b04dbe9e6bc72b0a9bfa