cr50: Only drive CCD_MODE_L when in CCD mode.
This commit changes the behaviour of the CCD_MODE_L pin.  When Cr50 is
not in CCD mode, it will stop driving the pin and turn it into an input.
This allows the pin to be driven elsewhere.  However, when Cr50 decides
to enable CCD mode, it will drive the pin low.

BUG=b:35804738
BRANCH=None
TEST=Flash dev board, use `ccd` console command to both enable and
disable CCD.  Verify that states are stable.  Verify that when CCD is
enabled, the state of DIOM1 does not disable CCD.  Verify that when CCD
is disabled, pulling DIOM1 low enables CCD.  Letting it float disables
CCD.
TEST=Verify that CCD mode is reflected in the device state.

Change-Id: I44645f28b362977ca6a502b646e4f4ff1a7430c7
Signed-off-by: Aseda Aboagye <aaboagye@google.com>